+++ client/src/components/Body.js
@@ -6,6 +6,8 @@ import PrivateRoute from './privateRoute';
 import ItemInfo from './ItemInfo';
 import CreateItem from './CreateItem';
 import EditItem from './EditItem';
+import AdminPage from './AdminPage';
+import EditCollection from './EditCollection';
 
 function Body() {
 
@@ -18,6 +20,8 @@ function Body() {
           <PrivateRoute exact path="/user/:id" component={UserPage} />
           <PrivateRoute exact path="/collection/:id" component={CreateItem} />
           <PrivateRoute exact path="/editItem/:id" component={EditItem} />
+          <PrivateRoute exact path="/admin/:id" component={AdminPage} />
+          <PrivateRoute exact path="/editCollection/:id" component={EditCollection} />
         </Switch>
       </BrowserRouter>
     </div>