+++ client/src/components/TableCollection.js
@@ -2,11 +2,12 @@ import React from 'react';
 import { Card } from 'react-bootstrap';
 import {  MarkdownPreview  } from 'react-marked-markdown';
 import { Link } from 'react-router-dom';
+import DeleteButtonCollect from './DeleteButtonCollect';
 
 export default function TableCollection(props) {
     const { dataCollect } = props;
 
-    const makeCollection = dataCollect.map((el, idx) => {
+    const makeCollection = dataCollect.reverse().map((el, idx) => {
         return (
             <Card style={{ width: '15rem' }} key={idx}>
                 <Card.Img variant="top" src={el.url} />
@@ -15,6 +16,8 @@ export default function TableCollection(props) {
                     <Card.Text>{el.collections}</Card.Text>
                     <MarkdownPreview value={el.description} />
                     <Link to={`/collection/${el._id}`}> Посмотреть </Link>
+                    <Link to={`/editCollection/${el._id}`}> Редактировать </Link>
+                    <DeleteButtonCollect id={el._id} />
                 </Card.Body>
             </Card>
         )