+++ client/src/components/UserPage.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import CreateCollection from './CreateCollection';
 import { useAuth0 } from "@auth0/auth0-react";
 
-function UserPage() {
+function UserPage(props) {
   const { isAuthenticated } = useAuth0();
   return (
     isAuthenticated && (
@@ -12,7 +12,7 @@ function UserPage() {
               <Link to='/'>Перейти на главную страницу</Link>
           </div>
           <div className='body-user'>
-              <CreateCollection/>
+              <CreateCollection idLink={props.location.pathname.slice(6)} />
           </div>
       </div>
     )